Remove magic_enum dependency to reduce build times
Replace all magic_enum usages (~184 across 87 files) with manual
enumToString/fromString functions per enum type. The generic
fmt::formatter in EnumReflection.h now uses a has_enumToString concept
with ADL lookup, so all existing fmt::format/LOG_* calls continue to
work. For nested class enums and third-party enums where ADL cannot
reach, static_cast<int>() is used at call sites.

This eliminates ~630s of header parse time (across 3,561 TUs) and
~507s of template instantiation time from the build.

+++ src/Backups/BackupCoordinationStageSync.cpp
@@ -437,7 +437,7 @@ void BackupCoordinationStageSync::readInitiatorVersion()
         String initiator_start_node;
         if (!zookeeper->tryGet(initiator_start_node_path, initiator_start_node))
         {
-            LOG_TRACE(log, "Couldn't read the initiator's version, assuming {}", kInitialVersion);
+            LOG_TRACE(log, "Couldn't read the initiator's version, assuming {}", static_cast<int>(kInitialVersion));
         }
         std::lock_guard lock{mutex};
         state.hosts.at(String{kInitiator}).version = parseStartNode(initiator_start_node, String{kInitiator});
