+++ src/DataTypes/Serializations/SerializationSubObjectSharedData.cpp
@@ -98,7 +98,7 @@ void SerializationSubObjectSharedData::enumerateStreams(
         else
         {
             /// If we add new serialization version in future and forget to implement something, better to get an exception instead of doing nothing.
-            throw Exception(ErrorCodes::NOT_IMPLEMENTED, "enumerateStreams is not implemented for shared data serialization version {}", serialization_version.value);
+            throw Exception(ErrorCodes::NOT_IMPLEMENTED, "enumerateStreams is not implemented for shared data serialization version {}", static_cast<int>(serialization_version.value));
         }
 
         settings.path.pop_back();
@@ -168,7 +168,7 @@ void SerializationSubObjectSharedData::deserializeBinaryBulkStatePrefix(
     else
     {
         /// If we add new serialization version in future and forget to implement something, better to get an exception instead of doing nothing.
-        throw Exception(ErrorCodes::NOT_IMPLEMENTED, "deserializeBinaryBulkStatePrefix is not implemented for shared data serialization version {}", serialization_version.value);
+        throw Exception(ErrorCodes::NOT_IMPLEMENTED, "deserializeBinaryBulkStatePrefix is not implemented for shared data serialization version {}", static_cast<int>(serialization_version.value));
     }
 
     state = std::move(shared_data_state);
@@ -329,7 +329,7 @@ void SerializationSubObjectSharedData::deserializeBinaryBulkWithMultipleStreams(
     else
     {
         /// If we add new serialization version in future and forget to implement something, better to get an exception instead of doing nothing.
-        throw Exception(ErrorCodes::NOT_IMPLEMENTED, "deserializeBinaryBulkWithMultipleStreams is not implemented for shared data serialization version {}", serialization_version.value);
+        throw Exception(ErrorCodes::NOT_IMPLEMENTED, "deserializeBinaryBulkWithMultipleStreams is not implemented for shared data serialization version {}", static_cast<int>(serialization_version.value));
     }
 }
