+++ src/Storages/MergeTree/MergeTreeIndexConditionText.cpp
@@ -332,7 +332,7 @@ bool MergeTreeIndexConditionText::mayBeTrueOnGranule(MergeTreeIndexGranulePtr id
         }
         else
         {
-            throw Exception(ErrorCodes::LOGICAL_ERROR, "Unexpected function type {} in MergeTreeIndexConditionText::RPNElement", element.function);
+            throw Exception(ErrorCodes::LOGICAL_ERROR, "Unexpected function type {} in MergeTreeIndexConditionText::RPNElement", static_cast<int>(element.function));
         }
 
         if (update_partial_disjunction_result_fn)
